Refactored LoginService
