+++ mobile/src/main/java/com/wedriveu/mobile/service/login/LoginServiceImpl.java
@@ -1,6 +1,4 @@
-package com.wedriveu.mobile.login.service;
-
-import com.wedriveu.mobile.model.User;
+package com.wedriveu.mobile.service.login;
 
 /**
  * Created by Marco on 12/07/2017.
@@ -9,14 +7,8 @@ public class LoginServiceImpl implements LoginService {
 
     @Override
     public void login(String username, String password, LoginServiceCallback callback) {
-        //TODO Complete with retrofit http call
-        if(username.isEmpty() || password.isEmpty()) {
-            callback.onLoginFinished(null, "Error");
-        }
-        if(username.equals("pinco") && password.equals("pallo")) {
-            User user = new User(username, password);
-            callback.onLoginFinished(user, null);
-        }
+        //TODO Complete with retrofit http request
+        callback.onLoginFinished(null, "Error");
     }
 
 }
